Patch to continue training where if was paused
For some reason, command line flags are not picked up by the program (or at least that is my theory for now that I have not tested). When loading a program, all flags are overwritten by the original ones when the model was initially trained. Because no new flags are picked up, the flags (that I want to set in the config.py file) is overwritten as well, not allowing me to load the model. To fix it, I just manually exclude the load_dir flag from getting overwritten. Way too long story for 1 line of code, oh well